requirements.txt
alembic==1.13.3
amqp==5.2.0
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
cssbeautifier==1.15.1
djlint==1.35.2
dnspython==2.6.1
EditorConfig==0.12.4
email_validator==2.2.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsbeautifier==1.15.1
json5==0.9.25
kombu==5.4.2
logging==0.4.9.6
Mako==1.3.5
MarkupSafe==2.1.5
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyupbit==0.2.34
PyYAML==6.0.2
redis==5.1.1
regex==2024.9.11
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.35
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.4
WTForms==3.1.2
